requirements.txt
adblockparser==0.5
argh==0.26.2
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
blinker==1.4
certifi==2016.2.28
cffi==1.6.0
click==6.6
ConfigArgParse==0.10.0
construct==2.5.2
cryptography==1.3.2
enum34==1.1.6
Flask==0.10.1
h2==2.3.1
hpack==2.2.0
html2text==2016.4.2
hyperframe==3.2.0
idna==2.1
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
lxml==3.6.0
MarkupSafe==0.23
mitmproxy==0.17
passlib==1.6.5
pathtools==0.1.2
Pillow==3.2.0
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==16.0.0
pyparsing==2.1.4
pyperclip==1.5.27
PyYAML==3.11
re2==0.2.22
requests==2.9.2
singledispatch==3.4.0.3
six==1.10.0
tornado==4.3
urwid==1.3.1
watchdog==0.8.3
Werkzeug==0.11.9
